chore: move fakeValidator from individual file into Validator_test.go
